Move URL stuff into Location from Diagnostic
