+++ compiler/hie-core/src/Development/IDE/LSP/LanguageServer.hs
@@ -19,6 +19,7 @@ import qualified Development.IDE.LSP.Definition as LS.Definition
 import qualified Development.IDE.LSP.Hover      as LS.Hover
 import qualified Development.IDE.Logger as Logger
 import Development.IDE.State.Service
+import Development.IDE.Types.Location
 
 import qualified Data.Aeson                                as Aeson
 import qualified Data.Rope.UTF16 as Rope
@@ -26,7 +27,6 @@ import qualified Data.Set                                  as S
 import qualified Data.Text as T
 
 import Development.IDE.State.FileStore
-import Development.IDE.Types.Diagnostics
 
 import qualified Network.URI                               as URI
